Split serverloop into sync and async version
