+++ examples/cdc_acm_serial.rs
@@ -3,13 +3,15 @@ use std::net::*;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
+use usbip::server::*;
+
 #[tokio::main]
 async fn main() {
     env_logger::init();
     let handler =
         Arc::new(Mutex::new(Box::new(usbip::cdc::UsbCdcAcmHandler::new())
             as Box<dyn usbip::UsbInterfaceHandler + Send>));
-    let server = Arc::new(usbip::UsbIpServer::new_simulated(vec![
+    let server = SyncUsbIpServer::new_simulated(vec![
         usbip::UsbDevice::new(0).with_interface(
             usbip::ClassCode::CDC as u8,
             usbip::cdc::CDC_ACM_SUBCLASS,
@@ -18,9 +20,9 @@ async fn main() {
             usbip::cdc::UsbCdcAcmHandler::endpoints(),
             handler.clone(),
         ),
-    ]));
+    ]);
     let addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), 3240);
-    tokio::spawn(usbip::server(addr, server));
+    tokio::spawn(server.serve(addr));
 
     loop {
         // sleep 1s
