+++ examples/hid_keyboard.rs
@@ -3,6 +3,8 @@ use std::net::*;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
+use usbip::server::*;
+
 #[tokio::main]
 async fn main() {
     env_logger::init();
@@ -10,7 +12,7 @@ async fn main() {
         Box::new(usbip::hid::UsbHidKeyboardHandler::new_keyboard())
             as Box<dyn usbip::UsbInterfaceHandler + Send>,
     ));
-    let server = Arc::new(usbip::UsbIpServer::new_simulated(vec![
+    let server = SyncUsbIpServer::new_simulated(vec![
         usbip::UsbDevice::new(0).with_interface(
             usbip::ClassCode::HID as u8,
             0x00,
@@ -24,9 +26,9 @@ async fn main() {
             }],
             handler.clone(),
         ),
-    ]));
+    ]);
     let addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), 3240);
-    tokio::spawn(usbip::server(addr, server));
+    tokio::spawn(server.serve(addr));
 
     loop {
         // sleep 1s
