+++ examples/host.rs
@@ -1,13 +1,15 @@
 use std::net::*;
-use std::sync::Arc;
+
 use std::time::Duration;
+
+use usbip::server::*;
 
 #[tokio::main]
 async fn main() {
     env_logger::init();
-    let server = Arc::new(usbip::UsbIpServer::new_from_host());
+    let server = usbip::server::SyncUsbIpServer::new_from_host();
     let addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), 3240);
-    tokio::spawn(usbip::server(addr, server));
+    tokio::spawn(server.serve(addr));
 
     loop {
         // sleep 1s
