+++ src/usbip_protocol.rs
@@ -346,7 +346,7 @@ impl UsbIpResponse {
                 let mut result =
                     Vec::with_capacity(48 + transfer_buffer.len() + iso_packet_descriptor.len());
 
-                debug_assert!(header.command == USBIP_RET_SUBMIT.into());
+                debug_assert!(header.command == <u16 as Into<u32>>::into(USBIP_RET_SUBMIT.into()));
 
                 result.extend_from_slice(&header.to_bytes());
                 result.extend_from_slice(&status.to_be_bytes());
@@ -362,7 +362,7 @@ impl UsbIpResponse {
             Self::UsbIpRetUnlink { ref header, status } => {
                 let mut result = Vec::with_capacity(48);
 
-                debug_assert!(header.command == USBIP_RET_UNLINK.into());
+                debug_assert!(header.command == <u16 as Into<u32>>::into(USBIP_RET_UNLINK.into()));
 
                 result.extend_from_slice(&header.to_bytes());
                 result.extend_from_slice(&status.to_be_bytes());
